@cuviper 2019-02-14
Before adding a new separate method, I think we can improve on the normal operator.

> Right now all implementations of Rem return a `BigUint` and are somewhat slow.

The output is one consideration, but I think the bigger issue for performance is that we're always computing the quotient and remainder together, just to return the latter.

https://github.com/rust-num/num-bigint/blob/eb003cc41b604658c885c052583ea9f453f9c910/src/biguint.rs#L1010-L1018

Currently we always need `self` by value (requiring a clone if we start with LHS as a reference), because `div_rem_digit` will write the quotient in place. I think we could change this to a simpler `rem_digit` which takes the LHS by reference, and doesn't need to build the quotient at all.

> Due to the fact that `a % b < b` is always true, we might want to change Rem to return a value of type `b` instead of a.

That's a reasonable possibility, but it would be a breaking change.